requirements.txt
django
djangorestframework
markdown
django-filter
pygments
coreapi
django-cors-headers